Solve 2021 day 8
This is a brute-ish force solution. I solved the problem in a playground this morning, and I'll
leave this for optimisation later.

Parsing could be better, so I don't do all the parsing for both problems.

Some kind of lazy / dynamic solution of only the numbers that are needed, rather than finding all
the mappings.

There should be more type safety around the mapping and turning an array of numbers into a digit,
it's a bit too dependent on assuming that only single digits are used.